+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Allow image uploads in shared excalidraw notes
 * Markdown editor: Prevent browser from opening some links in markdown preview
 * Rank project search results to prioritize name and tag matches over content
+* Plugin `jira`: Follow API pagination to list all available projects
 
 
 ## v2026.68 - 2026-08-19